refactor(api): move saved views to feature module layout
