+++ apps/api/src/modules/share/index.ts
@@ -5,7 +5,7 @@ import { authContext } from '#shared/auth-context';
 import { HttpError } from '#shared/lib';
 import { commonErrors, errors } from '#shared/responses';
 import { getIssueProjectId } from '../../issues/store';
-import { getView } from '../../views/store';
+import { getView } from '#modules/views/service';
 import {
   BundleResponse,
   ShareTokenResponse,
